requirements.txt
# Production dependencies
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
httpx>=0.25.0
pymupdf>=1.23.0
uvicorn>=0.24.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

